BNB 4-bit quantization support
