+++ lit_gpt/config.py
@@ -1,4 +1,3 @@
-import math
 from dataclasses import dataclass
 from typing import Optional, Any, Type, Literal
